fix: smooth works again, and NMR parameter upgrade

+++ silq/analysis/analysis.py
@@ -68,7 +68,7 @@ def smooth(x, window_len=11, window='hanning'):
         w = eval('np.' + window + '(window_len)')
 
     y = np.convolve(w / w.sum(), s, mode='valid')
-    return y[(window_len/2-1):-(window_len/2)]
+    return y[int(window_len/2-1):-int(window_len/2)]
 
 
 def find_high_low(traces, plot=False, threshold_peak=0.02, attempts=8,
